Allow updates to existing winget manifests
We recently received notification from a customer that installing
`microsoft/git` via winget was broken due to a signing issue. This happened
because we accidentally deployed an incorrect SHA for the `microsoft/git`
version `2.33.0.0.0` manifest on `microsoft/winget-pkgs` then replaced
the asset on GitHub with a new version with a new SHA.

When we tried to re-run the `update-winget` workflow
to correct the SHA, we discovered some issues with the action that make it
impossible to update existing manifests. This PR updates the action to allow
us to both create new winget manifests and update existing ones.

[Here](https://github.com/ldennington/winget-playground/pull/47) is an
example of a test PR opened to update an existing manifest with these
changes. And
[here](https://github.com/ldennington/winget-playground/pull/48) is an
example of a test PR that creates a new manifest with these changes.